requirements.txt
confluent-kafka==2.6.1
minio==7.2.12
psycopg2-binary==2.9.10
python-dotenv==1.0.1